requirements.txt
praw>=7.7
yt-dlp>=2024.1.0
google-api-python-client>=2.100
google-auth>=2.23
google-auth-oauthlib>=1.1
python-dotenv>=1.0
APScheduler>=3.10
flask>=3.0
